Add shared test fixtures
